imgproc/requirements.txt
pillow==11.3.0
opencv-python==4.12.0.88
opencv-contrib-python==4.12.0.88
pdf2image==1.17.0
numpy==2.0.2